Created Info.txt
Keep in mind this is a placeholder where you would literally just paste in your client secret. Are there better ways to handle this than a txt file handler? Hell yeah. Will I be going over any of that? Hell no. It's just straight-up easier to do this so that if anyone wants to do a similar project they can just jump in quickly and I also don't have to share my secret with others lol